Refactor of filter system
Should do the same thing as before but probably more keen to bugs. `ItemBill.cs` can also be implemented in `MainPageViewModel.cs` by replacing standard Bills. The advantage of this is that property `IsSelected` in `Bill` will not be needed anymore (cleaner and more compact database).

In sorting (or advance search) new Command will be needed for editing bills, this feature is not implemeted yet.